requirement.txt
asn1crypto==0.24.0
Babel==2.5.3
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
Django==2.0.3
django-bootstrap-form==3.3
django-debug-toolbar==1.9.1
django-formtools==2.1
django-otp==0.4.3
django-phonenumber-field==1.3.0
django-two-factor-auth==1.7.0
idna==2.6
ndg-httpsclient==0.4.4
phonenumberslite==8.9.1
Pillow==5.0.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.6.0
pyOpenSSL==17.5.0
PySocks==1.6.8
pytz==2018.3
qrcode==4.0.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
twilio==6.10.5
urllib3==1.22
# visit these listed sites to install certain libraries
# https://django-user-sessions.readthedocs.io/en/stable/installation.html
# https://docs.djangoproject.com/en/2.0/ref/contrib/gis/geoip2/
# coinmarketcap
# simplejson